rename file, reuse code
